Move accumulate function into separate mod
